refactor(poll): PollMainPage 컴포넌트 추상화 조정
